bt_generator/requirements.txt
torch
accelerate
bitsandbytes==0.41.0
peft
transformers>=4.31.0
sentencepiece
scp
paramiko
pyyaml